Move `Mode` out of `utils`
As is very much domain logic
