+++ src/entity.rs
@@ -2,6 +2,35 @@
 use crate::ansi;
 use crate::config;
 use crate::utils;
+use std::str::FromStr;
+
+// -------
+// SYMBOLS
+// -------
+
+//  The character set to use for the entities
+#[derive(Clone, Copy, Debug)]
+#[allow(dead_code)]
+pub enum Mode {
+    Original, //  Katakana
+    Binary,   //  0 or 1
+    ASCII,    //  ASCII
+    Braille,  //  Braille
+}
+
+// Implement the `FromStr` trait for Mode so that we can parse the command-line argument
+impl FromStr for Mode {
+    type Err = anyhow::Error;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_str() {
+            "original" => Ok(Mode::Original),
+            "binary" => Ok(Mode::Binary),
+            "ascii" => Ok(Mode::ASCII),
+            "braille" => Ok(Mode::Braille),
+            _ => Err(anyhow::Error::msg("Invalid Mode")),
+        }
+    }
+}
 
 //  ======
 //  ENTITY
@@ -20,7 +49,7 @@ pub struct Entity {
     /// entity symbol
     symbol: char,
     /// character set mode
-    mode: utils::Mode,
+    mode: Mode,
     /// frame-count since last switch
     frame_count: u16,
     /// number of frames before character switch
@@ -50,25 +79,25 @@ impl Entity {
     pub fn set_symbol(&mut self) {
         match self.mode {
             //  Katakana Symbols
-            utils::Mode::Original => {
+            Mode::Original => {
                 let r = utils::random_between(0x30a0, 0x30a0 + 96) as u32;
                 self.symbol = std::char::from_u32(r).unwrap_or('0');
             }
 
             //  Binary Symbols
-            utils::Mode::Binary => {
+            Mode::Binary => {
                 let r = utils::random_between(0, 2);
                 self.symbol = if r == 0 { '0' } else { '1' };
             }
 
             //  ASCII Symbols
-            utils::Mode::ASCII => {
+            Mode::ASCII => {
                 let r = utils::random_between(33, 127) as u32;
                 self.symbol = std::char::from_u32(r).unwrap_or('0');
             }
 
             //  Braille Symbols
-            utils::Mode::Braille => {
+            Mode::Braille => {
                 let r = utils::random_between(0x2840, 0x2840 + 63) as u32;
                 self.symbol = std::char::from_u32(r).unwrap_or('0');
             }
@@ -107,10 +136,10 @@ impl Entity {
             ansi::cursor_move_to(last_y as u32, self.x as u32);
         }
         match self.mode {
-            utils::Mode::Original => print!("  "),
-            utils::Mode::ASCII => print!(" "),
-            utils::Mode::Binary => print!(" "),
-            utils::Mode::Braille => print!(" "),
+            Mode::Original => print!("  "),
+            Mode::ASCII => print!(" "),
+            Mode::Binary => print!(" "),
+            Mode::Braille => print!(" "),
         }
     }
 }
