+++ src/utils.rs
@@ -1,35 +1,9 @@
 //  Library
 use rand::Rng;
-use std::str::FromStr;
 
 //  =========
 //  UTILITIES
 //  =========
-
-//  Matrix mode determines the character set to use for the entities
-#[derive(Clone, Copy, Debug)]
-#[allow(dead_code)]
-pub enum Mode {
-    Original,
-    Binary,
-    ASCII,
-    Braille,
-}
-
-// Implement the FromStr trait for Mode to parse the command-line argument
-impl FromStr for Mode {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "original" => Ok(Mode::Original),
-            "binary" => Ok(Mode::Binary),
-            "ascii" => Ok(Mode::ASCII),
-            "braille" => Ok(Mode::Braille),
-            _ => Err(anyhow::Error::msg("Invalid Mode")),
-        }
-    }
-}
 
 /// Generate a random number between min and max
 pub fn random_between<T: PartialOrd + rand::distributions::uniform::SampleUniform>(
@@ -64,6 +38,7 @@ mod tests {
     fn type_of<T>(_: &T) -> &str {
         std::any::type_name::<T>()
     }
+
     #[test]
     fn random_i32s_return_i32() {
         let number = super::random_between(-15, 32);
